requirements.txt
# YOLO and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0

# PyTorch with CUDA support for GPU training
torch>=2.0.0
torchvision>=0.15.0

# Additional ML/Data Science
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
seaborn>=0.12.0

# GUI and Display
cvzone>=1.5.6

# Progress tracking
tqdm>=4.65.0

# YAML processing
PyYAML>=6.0

# Image processing
imageio>=2.28.0

# Additional utilities
requests>=2.31.0
psutil>=5.9.0

# RKNN Toolkit for RDK X5 NPU
# Note: Install from official Rockchip repository
# git clone https://github.com/rockchip-linux/rknn-toolkit2.git
# cd rknn-toolkit2/rknn-toolkit2/packages
# pip install rknn_toolkit2-*-py3-none-any.whl

# Alternative ONNX support (if RKNN models unavailable)
onnx>=1.14.0
onnxruntime>=1.15.0

# Linux Audio Dependencies
# Install system packages:
# Ubuntu/Debian: sudo apt-get install espeak espeak-data alsa-utils pulseaudio
# CentOS/RHEL: sudo yum install espeak alsa-utils pulseaudio
# Arch Linux: sudo pacman -S espeak alsa-utils pulseaudio
